services/ingestor/requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.0.1
psycopg[binary]==3.2.5